+++ packages/devtools_app/lib/src/shared/analytics/constants/_vs_code_sidebar_constants.dart
@@ -14,6 +14,11 @@ enum VsCodeFlutterSidebar {
 
   static String get id => VsCodeFlutterSidebar.vsCodeFlutterSidebar.name;
 
+  /// Analytics event for when a request to enable a new platform type is sent
+  /// to VS Code.
+  static String enablePlatformType(String platformType) =>
+      'enablePlatformType-$platformType';
+
   /// Analytics event that is sent when a DevTools screen is opened from the
   /// actions toolbar for a debug session.
   static String openDevToolsScreen(String screen) =>