+++ packages/devtools_app/lib/src/shared/config_specific/logger/_allowed_error_web.dart
@@ -2,8 +2,9 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// ignore: avoid_web_libraries_in_flutter, as designed
-import 'dart:html';
+import 'dart:js_interop';
+
+import 'package:web/helpers.dart';
 
 /// Catch and print errors from the given future. These errors are part of
 /// normal operation for an app, and don't need to be reported to analytics
@@ -12,10 +13,9 @@ Future<T> allowedError<T>(Future<T> future, {bool logError = true}) {
   return future.catchError((Object error) {
     if (logError) {
       final errorLines = error.toString().split('\n');
-      window.console
-          .groupCollapsed('[${error.runtimeType}] ${errorLines.first}');
-      window.console.log(errorLines.skip(1).join('\n'));
-      window.console.groupEnd();
+      console.groupCollapsed('[${error.runtimeType}] ${errorLines.first}'.toJS);
+      console.log(errorLines.skip(1).join('\n').toJS);
+      console.groupEnd();
     }
     return Future<T>.value();
   });