+++ packages/devtools_app/lib/src/standalone_ui/api/impl/vs_code_api.dart
@@ -60,6 +60,14 @@ final class VsCodeApiImpl extends ToolApiImpl implements VsCodeApi {
   }
 
   @override
+  Future<bool> enablePlatformType(String platformType) {
+    return sendRequest(
+      VsCodeApi.jsonEnablePlatformTypeMethod,
+      {VsCodeApi.jsonPlatformTypeParameter: platformType},
+    );
+  }
+
+  @override
   Future<void> openDevToolsPage(String debugSessionId, String page) {
     return sendRequest(
       VsCodeApi.jsonOpenDevToolsPageMethod,
@@ -159,6 +167,7 @@ class VsCodeDebugSessionImpl implements VsCodeDebugSession {
     required this.flutterMode,
     required this.flutterDeviceId,
     required this.debuggerType,
+    required this.projectRootPath,
   });
 
   VsCodeDebugSessionImpl.fromJson(Map<String, Object?> json)
@@ -172,6 +181,8 @@ class VsCodeDebugSessionImpl implements VsCodeDebugSession {
               json[VsCodeDebugSession.jsonFlutterDeviceIdField] as String?,
           debuggerType:
               json[VsCodeDebugSession.jsonDebuggerTypeField] as String?,
+          projectRootPath:
+              json[VsCodeDebugSession.jsonProjectRootPathField] as String?,
         );
 
   @override
@@ -192,6 +203,9 @@ class VsCodeDebugSessionImpl implements VsCodeDebugSession {
   @override
   final String? debuggerType;
 
+  @override
+  final String? projectRootPath;
+
   Map<String, Object?> toJson() => {
         VsCodeDebugSession.jsonIdField: id,
         VsCodeDebugSession.jsonNameField: name,
@@ -199,6 +213,7 @@ class VsCodeDebugSessionImpl implements VsCodeDebugSession {
         VsCodeDebugSession.jsonFlutterModeField: flutterMode,
         VsCodeDebugSession.jsonFlutterDeviceIdField: flutterDeviceId,
         VsCodeDebugSession.jsonDebuggerTypeField: debuggerType,
+        VsCodeDebugSession.jsonProjectRootPathField: projectRootPath,
       };
 }
 
@@ -206,6 +221,7 @@ class VsCodeDevicesEventImpl implements VsCodeDevicesEvent {
   VsCodeDevicesEventImpl({
     required this.selectedDeviceId,
     required this.devices,
+    required this.unsupportedDevices,
   });
 
   VsCodeDevicesEventImpl.fromJson(Map<String, Object?> json)
@@ -216,6 +232,11 @@ class VsCodeDevicesEventImpl implements VsCodeDevicesEvent {
               .map((item) => Map<String, Object?>.from(item))
               .map((map) => VsCodeDeviceImpl.fromJson(map))
               .toList(),
+          unsupportedDevices:
+              (json[VsCodeDevicesEvent.jsonUnsupportedDevicesField] as List?)
+                  ?.map((item) => Map<String, Object?>.from(item))
+                  .map((map) => VsCodeDeviceImpl.fromJson(map))
+                  .toList(),
         );
 
   @override
@@ -224,9 +245,13 @@ class VsCodeDevicesEventImpl implements VsCodeDevicesEvent {
   @override
   final List<VsCodeDevice> devices;
 
+  @override
+  final List<VsCodeDevice>? unsupportedDevices;
+
   Map<String, Object?> toJson() => {
         VsCodeDevicesEvent.jsonSelectedDeviceIdField: selectedDeviceId,
         VsCodeDevicesEvent.jsonDevicesField: devices,
+        VsCodeDevicesEvent.jsonUnsupportedDevicesField: unsupportedDevices,
       };
 }